requirements.txt
discord.py[speed]
jsonpickle
pymongo

# Linter
pylint
# Formatter
autopep8
